CurrentWeekNav functionality complete & backend updates
The CurrentWeekNav now updates the ExercisesForDay when a new day is clicked.

Added a Metric collection-type to the backend.  Every profile will be able to choose a default metric - kg or lbs.  A relationship was created "Profile has one Metric".

Reorganized some components on the backend.  Still a work in progress, but I think the backend table structure is pretty close to how I want it.