Created search bar and added alt tags
The work on the search bar to find and add exercises to a user's schedule/calendar is underway. Realized I forgot to add alt tags to several img elements, so fixed that.

Also started work on the authenticated nav header.  Right now, I'm displaying the public header, there will be a different header for logged in users.