Implemented authenticated nav & progress on add exercise modal
The layout now displays either the public "Header" or "AuthenticatedNav" depending on the presence of a jwt (might consider renaming Header to PublicNav).

Tons of progress on the form to add exercise(s). The UI and form itself is complete.  Just need to set up the POST requests to send the values to the DB, then set up the GET requests (or might use GraphQL) to display the new data on the user's calendar.