Refactored AddExerciseFormik
AddExerciseFormik component was extremely long.  Refactored it into a dozen components, which also need to be refactored.

As of right now, there are components that perform the same tasks with different form values, which violateds DRY.  At some point I'm going to create 3 components that will allow me to delete 6 total components. First, I need to gain a better understanding of how to manage helper functions in Formik.