DB structure edits & Formik fixex
Decided to make a users program (exercises, sets, reps, etc.) separate from the profile (it was nested with a Strapi component) and give it its own content-type.  The nested data structure was a headache and the change allowed for shallower nesting.

Formik resets values to the initialValues when a component updates in the form.  This is not desirable, so for now I'm going to elevate the state (above the formik form 'AddExerciseFormik') and assign local values to the initialState where I need it to not reset.