Add exercise form progress
We now have it to where a successful POST of a program clears the Add Exercise form and closes the modal.  Added error messages to the required fields as well.  If a user tries to submit the form without providing a date, set, reps, or weight, the error messages appear underneath the fields. They also only show if error messages are present & if the fields have been touched.