Moved the Layout componet & updated Link components
I realized I could move the Layout component to the root _app.js since the layout being displayed is controlled by the jwt. Originally I was using different layouts depending on public vs private but that changed a while back.

Decided having a 'Back' button was a bit confusing given the calendar flow and switched it to a Link component that just goes back to the current year.

Made Links more dynamic. Passing the params prop to the Link as property is not sufficient since someone could navigate to a neighbor month in a DayView. Therefore, those Links needed to reflect that.