Made the [year] parameter dynamic & started WeekView
I had hardcoded the current year (2020) into all of my <Link> "as" attributes for convenience purposes.  Of course, a user needs to be able to switch years if they desire, as well as time moves forward and 2021 is coming last I checked.

I skipped over the WeekView a while back because I wasn't sure how I would design the UI.  Now I have a good idea and I need to insert it into the dynamic route system.  I'm going to simply take the first & last index of my week array and that will be the week range of dates passed to the URL.